renaming scaffold hook error
