+++ packages/nextjs/components/scaffold-eth/Contract/WriteOnlyFunctionForm.tsx
@@ -12,7 +12,7 @@ import {
   getParsedContractFunctionArgs,
   getParsedError,
 } from "~~/components/scaffold-eth";
-import { useTransactor } from "~~/hooks/scaffold-eth";
+import { useTransactor } from "~~/scaffoldHooks/scaffold-eth";
 import { getTargetNetwork, notification } from "~~/utils/scaffold-eth";
 
 type WriteOnlyFunctionFormProps = {
