+++ packages/nextjs/pages/_app.tsx
@@ -11,7 +11,7 @@ import { Header } from "~~/components/Header";
 import { BlockieAvatar } from "~~/components/scaffold-eth";
 import { RobotsProvider } from "~~/context/RobotsContext";
 import { Web3AuthProvider } from "~~/context/Web3AuthContext";
-import { useNativeCurrencyPrice } from "~~/hooks/scaffold-eth";
+import { useNativeCurrencyPrice } from "~~/scaffoldHooks/scaffold-eth";
 import { useGlobalState } from "~~/services/store/store";
 import { wagmiConfig } from "~~/services/web3/wagmiConfig";
 import { appChains } from "~~/services/web3/wagmiConnectors";
