+++ packages/nextjs/pages/subscribe/status/[planId].tsx
@@ -19,20 +19,20 @@ const StatusPage = () => {
   const [isLoading, setIsLoading] = useState(true);
   const router = useRouter();
 
-  const planId = router.query.planId as string;
+  const planID = router.query.planID as string;
 
   const redirectToSubscribeLanding = () => {
     router.push("/subscribe");
   };
 
   const browseToSubscriptionDetails = () => {
-    router.push(`/subscribe/${plan?.planId}`);
+    router.push(`/subscribe/${plan?.planID}`);
   };
 
   useEffect(() => {
     const checkUserOrdersForMatch = async () => {
       const orders = await getOrders();
-      const foundOrder = orders?.find(order => order.buyer === address && order.plan.planId === planId);
+      const foundOrder = orders?.find(order => order.buyer === address && order.plan.planID === planID);
       if (foundOrder) {
         setOrder(foundOrder);
       }
@@ -42,22 +42,22 @@ const StatusPage = () => {
     if (subsContract && address) {
       checkUserOrdersForMatch();
     }
-  }, [address, getOrders, planId, subsContract]);
+  }, [address, getOrders, planID, subsContract]);
 
   useEffect(() => {
     const setPlanIfExists = async () => {
       const plans = await getPlans();
-      const foundPlan = plans?.find(plan => plan.planId === planId);
+      const foundPlan = plans?.find(plan => plan.planID === planID);
 
       if (foundPlan) {
         setPlan(foundPlan);
       }
     };
 
-    if (subsContract && !plan && planId) {
+    if (subsContract && !plan && planID) {
       setPlanIfExists();
     }
-  }, [plan, planId, getPlans, subsContract]);
+  }, [plan, planID, getPlans, subsContract]);
 
   return (
     <div className="p-32 flex-grow" data-theme="exampleUi">
