+++ packages/nextjs/scaffoldHooks/scaffold-eth/index.ts
