+++ packages/nextjs/scaffoldHooks/scaffold-eth/useAccountBalance.ts
