+++ packages/nextjs/scaffoldHooks/scaffold-eth/useAnimationConfig.ts
