+++ packages/nextjs/scaffoldHooks/scaffold-eth/useAutoConnect.ts
