+++ packages/nextjs/scaffoldHooks/scaffold-eth/useBurnerWallet.ts
