+++ packages/nextjs/scaffoldHooks/scaffold-eth/useContractLogs.ts
