+++ packages/nextjs/scaffoldHooks/scaffold-eth/useDeployedContractInfo.ts
