+++ packages/nextjs/scaffoldHooks/scaffold-eth/useFetchBlocks.ts
