+++ packages/nextjs/scaffoldHooks/scaffold-eth/useNativeCurrencyPrice.ts
