+++ packages/nextjs/scaffoldHooks/scaffold-eth/useNetworkColor.ts
