+++ packages/nextjs/scaffoldHooks/scaffold-eth/useOutsideClick.ts
