+++ packages/nextjs/scaffoldHooks/scaffold-eth/useScaffoldEventHistory.ts
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { Abi, AbiEvent, ExtractAbiEventNames } from "abitype";
 import { Hash } from "viem";
 import { usePublicClient } from "wagmi";
-import { useDeployedContractInfo } from "~~/hooks/scaffold-eth";
+import { useDeployedContractInfo } from "~~/hooks/Botblock";
 import { replacer } from "~~/utils/scaffold-eth/common";
 import {
   ContractAbi,
