+++ packages/nextjs/types/typechain-types/factories/contracts/BotblockMarket__factory.ts
@@ -379,7 +379,7 @@ const _abi = [
     inputs: [
       {
         internalType: "uint256",
-        name: "planId",
+        name: "planID",
         type: "uint256",
       },
       {
